split base from index for core, cpu and webgl
