+++ tfjs-backend-cpu/src/index.ts
@@ -15,12 +15,6 @@
  * =============================================================================
  */
 
-import {registerBackend} from '@tensorflow/tfjs-core';
-import {MathBackendCPU} from './base';
-
-// Side effects for default initialization of MathBackendCPU
-registerBackend('cpu', () => new MathBackendCPU(), 1 /* priority */);
-import './register_all_kernels';
-
 // All exports from this package should be in base.
 export * from './base';
+import './register_all_kernels';
