+++ tfjs-backend-webgl/src/index.ts
@@ -15,18 +15,6 @@
  * =============================================================================
  */
 
-import {device_util, registerBackend} from '@tensorflow/tfjs-core';
-import {MathBackendWebGL} from './backend_webgl';
-export {version as version_webgl} from './version';
-
-if (device_util.isBrowser()) {
-  registerBackend('webgl', () => new MathBackendWebGL(), 2 /* priority */);
-}
+// All exports from this package should be in base.
+export * from './base';
 import './register_all_kernels';
-
-// Export webgl utilities
-export * from './webgl';
-
-// Export forceHalfFlost under webgl namespace for the union bundle.
-import {forceHalfFloat} from './webgl';
-export const webgl = {forceHalfFloat};
